+++ src/main/java/de/jeff_media/AngelChest/nbt/NBTTags.java
@@ -1,0 +1,5 @@
+package de.jeff_media.AngelChest.nbt;
+
+public class NBTTags {
+    public static final String IS_HOLOGRAM = "isHologram";
+}